Created models and mapped
Read the comment in app.js. Explained the working in respective files. Used the code given by the trainer that had ejs files. Learnt how to read ejs files but used the template given to save time as ejs is not used in the industry due to its slow nature.